+++ addon/utils.js
@@ -876,3 +876,17 @@ export async function fetchEntityDefinitions(addEntityCallback, apiIdentifier = 
     console.error("fetch entity definitions: ", err);
   });
 }
+
+/**
+ * Validates if a string is a valid Salesforce record ID
+ * @param {string} recordId - The string to validate
+ * @returns {boolean} True if the string is a valid record ID
+ */
+export function isRecordId(recordId) {
+  return typeof recordId === "string"
+       && /^[a-zA-Z0-9]{15,18}$/.test(recordId)
+       && /^[0-9a-zA-Z]{3}/.test(recordId)
+       && !recordId.startsWith("000")
+       && !/[^a-zA-Z0-9]/.test(recordId)
+       && /[0-9]/.test(recordId.slice(0, 5));
+}